add webpack alias & move css to directory
