+++ package.json
@@ -18,7 +18,7 @@
     "serialize-javascript": "^3.0.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-app-rewired start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -34,6 +34,7 @@
   ],
   "devDependencies": {
     "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2"
+    "@babel/core": "^7.2.2",
+    "react-app-rewired": "^2.1.6"
   }
 }
