--- src/component/hi.js
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-
-export class Template extends Component {
-    render() {
-        return (
-            <div> component </div>
-        );
-    }
-};
-export default Template;
