+++ src/css/Main.css
@@ -1,14 +1,16 @@
+:root {
+    --main-nav-color: #0e1219;
+    --main-text-color: rgba(240, 255, 255, 0.7);
+}
+
 * {
     margin: 0;
     font-family: 'Open Sans Condensed', sans-serif;
-    color:#6e7fa2;
+    color: var(--main-text-color);
 }
 
 .info-block {
     opacity: 0.6;
-    /* background-color: #6e7fa2; */
-    /* box-shadow: 0 0 3px 3px #28b9cb7; */
-    /* border-radius: 0.3rem; */
 }
 
 .content {
@@ -21,7 +23,7 @@
     width:6%;
     display: flex;
     flex-direction: column;
-    background-color: #0e1219;
+    background-color: var(--main-nav-color);
     justify-content: flex-start;
 
     transform-origin: left;
@@ -47,7 +49,6 @@
 .sidenav:hover a {
     transform-origin: center;
     text-align: center;
-    color: white;
     transform: scaleX(0.66);
     transition-duration: 0.2s;
 }
@@ -58,16 +59,37 @@
     font-size: 1.5rem;
     margin-left:auto;
     margin-right:auto;
-    color: azure;
-    opacity: 0.7;
     transition: text-shadow 0.3s, opacity 0.3s, color 0.3s;
 }
 
 .icon:hover {
-    color:greenyellow;
-    text-shadow: 0px 0px 30px greenyellow;
+    text-shadow: 0px 0px 30px;
     opacity: 1.5;
     transition: text-shadow 0.3s, opacity 0.3s, color 0.3s;
+}
+
+#home:hover {
+    color:greenyellow;
+}
+
+#map:hover {
+    color:#7DE2D1;
+}
+
+#music:hover {
+    color:greenyellow;
+}
+
+#art:hover {
+    color:#9797D4;
+}
+
+#weather:hover {
+    color:#9EBEFF;
+}
+
+#hobby:hover {
+    color:#FFABBC;
 }
 
 .sidenav-center-align {
@@ -103,15 +125,13 @@
 .page {
     width:100%;
     margin:6vh 0 0 5vw;
-    /* background-color: #9696AD */
-
 }
+
 .head h1{
     justify-self: left;
     letter-spacing: 2px;
     margin-top: 4vh;
     margin-left:1rem;
-    color: #9090a3;
 }
 .container{
     width:96%;
@@ -134,12 +154,8 @@
     width: 100%;
     padding-top: 5vh;
     text-align: center;
-    /* border-bottom: 2px solid #9090a3; */
-    /* background-color: aliceblue; */
 }
 .title h3{
-    color: #dedef9;
-    /* text-decoration: underline; */
     letter-spacing: 2px;
     z-index: 100;
 }
@@ -150,7 +166,6 @@
     /* min-width: 460px; */
     flex-direction: column;
     justify-content: space-evenly;
-    /* background-color: #9696AD;    */
 }
 
 .arc-chart{
