+++ src/routes/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Switch, Route, withRouter } from "react-router-dom";
 import { TransitionGroup, CSSTransition } from "react-transition-group";
 
-import Home from './component/pages/Home.jsx'
-import Map from './component/pages/Map.jsx'
+import Home from '@Component/pages/Home.jsx'
+import Map from '@Component/pages/Map.jsx'
 
 function Container ({location})  {
     return (
